Begins implementing Rubik's Cube.
At this point I realize that shoehorning Rubiks Cube into the existing Move pattern is a bad choice.
Moves should include direction and side in one object. Combining this with the Move enum used in NPuzzleState feels hacky.
Therefore, the next step is refactoring separate enums for NPuzzleMove and RubiksPuzzleMove.